requirements.txt
SQLAlchemy==2.0.27
fastapi==0.109.2
python-dotenv==1.0.1
psycopg2==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2